backend/requirements.txt
fastapi
uvicorn
python-multipart
tensorflow==2.20.0
numpy
pillow
h5py
